Define Gate Authorization

+++ app/Repositories/ReplySupportRepository.php
@@ -7,6 +7,7 @@ use App\Models\ReplySupport as ReplySupport;
 use App\Repositories\Contracts\ReplyRepositoryInterface;
 use Exception;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use stdClass;
 
 class ReplySupportRepository implements ReplyRepositoryInterface
@@ -41,6 +42,8 @@ class ReplySupportRepository implements ReplyRepositoryInterface
     public function delete(string $id): bool
     {
         if (! $reply = $this->model->find($id)) return false;
+        if (Gate::denies('owner', $reply->user->id)) abort(403, 'Not Authorized.');
+
         return (bool)$reply->delete();
     }
 }
